Removing fill attribute from ScrollTimeline
The spec work is still on going for this change, but this is a first
pass implementation that will hopefully help guide the specifics of the
spec.

A test has been added to verify that startScrollOffset and
endScrollOffset still affect animations as they should.

Current implementation returns -Infinity for when the timeline
current time is before startScrollOffset and Infinity for when
timeline current time is after endScrollOffset. This is likely to
change in the future after adding timeline states, but for the sake of
trying to keep this patch focused, that will be added in a later CL.

Bug: 1041621
Change-Id: I938679f0c61b42b588933bf9e5a8e7bd2776f529