[client hints] Fix handling of redirects in navigation
In particular, the browser needs to parse accept-ch there since
the renderer doesn't see the intermediate steps until after everything
is fetched, and new client hints may be requested in the middle of the
chain.

This is also likely the first step towards moving more of handling
of accept-ch to browser (which will simplify some things and fix
some bugs).

Bug: 1050726
Change-Id: Ie02d7492c02b75034b36af30e489ebbb430d6a1b